pyproject.toml
[project]
name = "renton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "dotenv-python>=0.0.1",
    "fake-useragent>=2.2.0",
    "loguru>=0.7.3",
    "pandas-stubs==2.3.2.250926",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.3",
    "utils>=1.0.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.14.3",
]
